requirements.txt
# Flask and web framework
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==2.3.7
gunicorn==21.2.0
python-dotenv==1.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Shopify integration
ShopifyAPI==12.2.0

# AI and ML dependencies
llama-index-core==0.10.10
llama-index-llms-nebius==0.1.3
llama-index-embeddings-nebius==0.1.2
llama-index-vector-stores-postgres==0.1.2
llama-index-readers-file==0.1.3
langchain==0.0.335
langchain-community==0.0.16
openai==1.3.7
transformers==4.35.2
torch==2.1.1
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
Pillow==10.1.0

# Web and API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
beautifulsoup4==4.12.2

# Utilities
python-slugify==8.0.1
email-validator==2.1.0.post1
pydantic==2.5.2
plotly==5.18.0
Markdown==3.5
